pyproject.toml
[project]
name = "seezeit-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.8",
    "langsmith>=0.4.32",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
]
